Sync changes with Upstream.
- Fixed a bug that might cause resource leak when an `acquire()` moved to the impossibleWaiters but then get its context canceled.

- Upstream changes introduced a change to the API Behavior in case of an acquire() that is bigger than the size of the semaphore.

Before: It was added to the queue (even though it was doomed to never get acquired)

Now: It returns an error.

In case of introducing Resize(), the old API behavior makes sense. because an acquire() is no longer 100% doomed to never get acquired. In case of a resize that increases the semaphore's size, the acquire can get unblocked.

I reverted the API this old behavior.