x/sync/singleflight/v2: add package
This adds a new package, x/sync/singleflight/v2, which is a version of
x/sync/singleflight that uses generics instead of `string`, and
`interface{}` in its API.

The biggest changes are to the `Do` methods, and the `Result` type. The
`Do` methods now take a `comparable` generic type instead of `string`
for its inputs (matching that of `map` and similar). The output type of
the `Do` method and the `Val` field in `Result` are now types that user
can specify instead `interface{}`.

Along the way, some tests received modifications to remove some
now-unneeded `fmt.Sprintf` calls or add an `empty` struct for nil return
tests.

Also, `ExampleGroup` also received some additions to make it clear that
non-`string` input types are acceptable.

This is following a similar pattern as discussed with the `math/rand/v2`
project. There is, however, one difference in affordances between
packages in the stdlib and outside of the stdlib that we try to
accomadate here.

Stdlib packages like `math/rand/v2` can rely on the Go compiler version
being the one our package is released with and, for packages outside of
the stdlib, the errors can sometimes be cryptic when the package needs a
more modern Go compile than a user attempted to use.

For instance, `singleflight/v2` has a build tag specifying the need for
Go 1.18 or later to be compiled in its necessary code files. When an
older compiler is used to build it, the user will get an error starting
with "build constraints exclude all Go files in". This makes sense when
you read all of the files, but a user may be using it as a transitive
dependency and won't know whether the build tags aren't matching because
their OS is unsupported or their CPU architecture or what.

To ameliorate user confusion, an extra `notgo18.go` file has been added
with build tags saying that it's only built if the Go compiler isn't
version 1.18 or later. And that file has a compile error in it that will
error like so:

    $  go1.17.2 build .
    # golang.org/x/sync/singleflight/v2
    ./notgo18.go:16:25: undefined: REQUIRES_GO_1_18_OR_LATER

That error message is an attempt to be more clear to users about what
needs to change in their build.

Another alternative to that would be to change x/sync's `go.mod` to
require Go 1.18 or greater. However, the rest of the x/sync packages
don't require Go 1.18, and that seems like too large of a breaking
change for singleflight/v2 alone.